stop using sql _compute
At the end of a sqlalchemy computation we need to wrap the result from compute
in a sqlalchemy.select statement.  We used to do this by having
compute(TableExpr, sqlalchemy.Table) call a specialized private _compute
function.  Now that compute/core.py has non-trivial functionality I think that
we need to connect up sql's private _compute dispatch system with compute.

To handle the need for the sqlalchemy.select call I've added an external,
non-dispatched computefull function in sqlalchemy.  I'm not yet sure how to
integrate this into the rest of the system.