ENH: kill the attr cache.
The attr cache held a reference to every blaze expression that ever had
an attribute looked up on it. This cache prevented them from ever
getting garbage collected.

Moving the attr_cache to the objects' __dict__ slot allows for the
object to become isolated and the cycle detector can kill the object.