BUG: fix inner_column optimizations
When we replace a select of 1 column with its inner column, we must
first verify that there are no other parts of the select that we are
missing. For example:

select table.column from table where column > 5

This select cannot be replaced with just table.column but

select table.column from table

can be. This optimization was used in some places, namely in BinOp
handling for the sql backend.