ENH: adds cast and coalesce
Cast is a type level function that reinterpretes a node as a different
type. This is like c cast, or c++ reinterpret_cast. This allows users to
correct mistakes in the blaze type system.

Coalesce is a binary operator much like sql coalesce. It picks the first
non null element.