Make Blaze server `/add` endpoint more flexible
Allow the `/add` endpoint of the Blaze server to take a more complex
payload. Rather than simply a name and resource URI string, the server
can now accept a dictionary with 'source', 'args, and 'kwargs' fields.

The 'source' is the URI string as before, but the 'args' field can be a
sequence of additional positional arguments, and the 'kwargs' fields can
be a dictionary of keyword arguments -- both are passed through to the
`blaze.resource` constructor.

For instance, in the test case the `encoding` argument is passed through
to allow reading a CSV in a non-default encoding.