Allow specifying modules that must be imported
Add an optional 'imports' key to the expanded payload dictionary, which
functions just like the 'imports' key in the server YAML file: specifies
string names for modules that must be imported before the resource can
be created.

In the ``blaze.server.client.Client``, this adds an optional keyword
argument to the `add` method -- while it is documented specifically, it
doesn't have any precedence over other kwargs.