Isolate `/add` endpoint behind a flag
Add a flag `allow_add` to the ``blaze.server.Server`` constructor to
determine whether or not the `/add` route is available -- it defaults to
`False`.

Currently, if this is not set, it doesn't define the URL mapping at all,
so any requests for it will result in a 404 Not Found error.