Accept more flexible content-type specification
Closes #1489. The `Content-Type` header can contain option information
in a semicolon-separated portion after the main content-type string. This
PR allows those more elaborate headers to be accepted by the server as
valid serialization formats.

To make this more extensible in future, it might be a good idea to use a
helper function such as that [available in
Werkzeug](http://werkzeug.pocoo.org/docs/0.11/http/#werkzeug.http.parse_options_header)
to handle this.