BUG: actually avoid the copy in fastmsgpack's deserialization
Why is bytes == int not a TypeError?

When we are decompressing the data from the server we can check if we
are the only owners the the bytes and if so just repurpose the buffer as
a mutable array. This is not 100% safe but it is probably fine. This is
also just a compat layer for older versions of pandas as this is handled
in a more safe way in pandasa 0.18.