Add debug logging for all events
Ensure that if desired logs can capture server startup and shutdown, as
well as detect when a dataset was added. Note that the compute operation
already has timing log output, and the default Flask logs for the
datashape check operation should contain all the information that was in
the request.