ENH: Add support for variadic expressions.
Adds a new VarArgsExpr/VarArgs pair of objects for representing
sequences of known types but unknown length in the expression tree. This
allows us to use multipledispatch to define compute handlers for
expressions that conceptually accept a sequence like `Merge`.

This also implements `Merge` as a variadic expression.

Future work will be porting `Concat`.