MAINT: Move expr and compute pieces out interactive.
Move the `_Data` class and rename it to `Data`.
(A soon to follow patch will add a `Literal` class from which `Data` will
inherit, and change dispatches to use `Literal`.)

Also, move compute and coerce functions into compute.core.

This is part of the effort of supporting a literal type which is importable from
expr instead of interactive.