ENH: Add `literal`, analogous to `data`.
The purpose of the `literal` function is to create leafs which delegate the repr to
the underlying data; e.g. `repr(frozenset([1, 2, 3]))` and
`repr(literal(frozenset([1, 2, 3]))` would be equivalent.

Change the `Data` class to `Literal`, and change the repr logic so that if
`_name` is `None` the repr of the `data` object is used.

Support for generated names which include the data type, dshape and name generated
by autoincrementing is retained by giving the `data` function a new default of
`name=generated`.

This patch also moves some tests from `test_interactive` to test suites
dedicated to `literal` and `data`.