ENH: Allow to_tree to render a literal.
So that input literals which are not bound to a data resource, can be sent to
the server. e.g. a frozenset which is the `_keys` of an `IsIn` statement.

Also, split literal and data so that one does not inherit from the other, to
make it easier to be precise with isinstance checks that target one but not the
other.

Co-Authored-By: <JoeJev@gmail.com>